casper: Move common.js and test-credentials.js to casper-lib/.
