+++ frontend_tests/casper_lib/.gitignore
