+++ frontend_tests/casper_tests/04-subscriptions.js
@@ -1,4 +1,4 @@
-var common = require('../common.js').common;
+var common = require('../casper_lib/common.js').common;
 
 common.start_and_log_in();
