+++ frontend_tests/casper_tests/07-stars.js
@@ -1,4 +1,4 @@
-var common = require('../common.js').common;
+var common = require('../casper_lib/common.js').common;
 
 function star_count() {
     return casper.evaluate(function () {
