+++ tools/generate-test-credentials
@@ -5,7 +5,7 @@ cd "$(dirname "$0")"/..
 email=iago@zulip.com
 
 password=$(python manage.py print_initial_password "$email" | fgrep "$email" | awk '{ print $2 }')
-cat > frontend_tests/casper_tests/test_credentials.js <<EOF
+cat > frontend_tests/casper_lib/test_credentials.js <<EOF
 // Generated by tools/generate-test-credentials
 var test_credentials = {default_user: {username: 'iago@zulip.com', password: '$password'}};
 try { exports.test_credentials = test_credentials; } catch (e) {}
